Adds code for multi-plot of variables of interest
All the bivariate situations are put together in this plot for an
overview. No titles or axis labels were included at this stage but the
maximum height on the y-axis is approximately 8,000 observations. The rows
are for outcomes stress, violence and bullying/harrassment, respectively.
At this point, the code is rudimentary and is just to allow the analyst
have an overview of the relationships. If it is decided that it will be
used beyond this point, then it will be updated to make it far more
communicative.