Creates new objects and adds housekeeping code
The two datasets i.e. training and test, are saved as distinct R objects
to enhance portability for use in the working on the data on different
devices and to provide to the opening of a new script file for regression
analysis.

Housekeeping code was also added, for whatever it's worth...