Adds a new function to enable multiple plots
I decided to spend a little time to build a function with which we can
easily plot the barcharts in such a manner that they can be compared
within each of the variable groups that were earlier created. This will
make the analysis easier to understand and easier to modify if needed.