Contains code for multivariate analysis proper
The code here was used as a test case for regression of variable
'concerns.stress' over predictors representiing the organisational
factors. This looks good so far and I will tidy it up in the next few
hours. The plan is for make it work as a plug-in for the other
outcome-predictor pairs in the research question.