Adds Formula object
I used a Formula object to enable me to do create a global formula for all
the variable pairs and this object was made subsettable for multiple
operations across the script. This makes it more readable and shortens the
code significantly. However, I observed that there may be need for some
optimization. The use of nested loops markedly increased processing time.